sync-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.10.0
apscheduler==3.10.4
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai==0.8.0
